Restructure integration tests
