+++ integration_tests/e2e/prisons/prisonStatus/prisonStatus.cy.ts
@@ -1,10 +1,10 @@
-import TestData from '../../../server/routes/testutils/testData'
-import HomePage from '../../pages/home'
-import Page from '../../pages/page'
-import SupportedPrisonsPage from '../../pages/prisons/prisonsPage'
-import ViewSessionTemplatesPage from '../../pages/prisons/viewSessionTemplatesPage'
-import PrisonStatusPage from '../../pages/prisons/prisonStatusPage'
-import { SessionTemplatesRangeType } from '../../../server/data/visitSchedulerApiTypes'
+import TestData from '../../../../server/routes/testutils/testData'
+import HomePage from '../../../pages/home'
+import Page from '../../../pages/page'
+import SupportedPrisonsPage from '../../../pages/prisons/prisonsPage'
+import ViewSessionTemplatesPage from '../../../pages/prisons/sessionTemplates/viewSessionTemplatesPage'
+import PrisonStatusPage from '../../../pages/prisons/prisonStatus/prisonStatusPage'
+import { SessionTemplatesRangeType } from '../../../../server/data/visitSchedulerApiTypes'
 
 context('Supported prisons', () => {
   const prisonCode = 'HEI'
