+++ integration_tests/pages/prisons/prisonStatus/prisonStatusPage.ts
@@ -1,4 +1,4 @@
-import Page, { PageElement } from '../page'
+import Page, { PageElement } from '../../page'
 
 export default class prisonStatusPage extends Page {
   constructor() {
