+++ integration_tests/pages/prisons/sessionTemplates/viewSessionTemplatePage.ts
@@ -1,4 +1,4 @@
-import Page, { PageElement } from '../page'
+import Page, { PageElement } from '../../page'
 
 export default class ViewSessionTemplatePage extends Page {
   constructor() {
