+++ integration_tests/pages/prisons/sessionTemplates/viewSessionTemplatesPage.ts
@@ -1,4 +1,4 @@
-import Page, { PageElement } from '../page'
+import Page, { PageElement } from '../../page'
 
 export default class ViewSessionTemplatesPage extends Page {
   constructor() {
